requirements.txt
requests
python-dotenv
openai
schedule
icalendar
# Pentru WhatsApp API (adaugă pachetul potrivit când integrezi)
